anoubisd fails to build with -Werror on amd64
Explicitly cast anoubis_cookie_t to unsigned long long in printfs.

Aegis-Change: anoubis.10.830.C127 Delta 9
Bug: 723
TE BLE

+++ src/libapn/apnparser.c
@@ -671,9 +671,9 @@ apn_print_scope(struct apn_scope *scope, FILE * file)
 	if (!scope)
 		return 0;
 	if (scope->task)
-		fprintf(file, " task %lld", scope->task);
+		fprintf(file, " task %llu", (unsigned long long)scope->task);
 	if (scope->timeout)
-		fprintf(file, " until %ld", (long int)scope->timeout);
+		fprintf(file, " until %lu", (long int)scope->timeout);
 	return 0;
 }
